transcript_processor/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
simple-salesforce
nltk